requirements.txt
gradio==3.38.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
opencv-python==4.8.0.76
numpy==1.24.3
gdown==4.7.1
pillow==10.0.0
git+https://github.com/facebookresearch/segment-anything.git